use double as Output type instead int + add formatting functions for test param names

+++ tasks/otcheskov_s_elem_vec_avg/common/include/common.hpp
@@ -8,8 +8,8 @@
 namespace otcheskov_s_elem_vec_avg {
 
 using InType = std::vector<int>;
-using OutType = int;
-using TestType = std::tuple<std::string, int>;
+using OutType = double;
+using TestType = std::tuple<std::string, double>;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace otcheskov_s_elem_vec_av
